telegram-bot/pyproject.toml
 [project]
 name = "amequizbot"
 version = "0.1.0"
 description = "Telegram bot to extract poll data from JSON messages."
 readme = "README.md"
 requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0",
  "python-dotenv>=1.0.0",
]
 
 [project.scripts]
 amequizbot = "amequizbot.main:main"
 
 [build-system]
 requires = ["setuptools>=68"]
 build-backend = "setuptools.build_meta"
